msm8937-common: check_features: fix NFC removal

+++ configs/check_features.sh
@@ -1,11 +1,10 @@
 #!/sbin/sh
 
-sku=`getprop ro.boot.hardware.sku`
 device=`getprop ro.boot.device`
 radio=`getprop ro.boot.radio`
 
-if [ "$device" == "cedric" ] || [ "$device" == "owens" ] || [ "$device" == "perry" ]; then
-    # owens, perry and cedric don't have e-compasses
+if [ "$device" != "cedric" ] && [ "$device" != "owens" ] && [ "$device" != "perry" ]; then
+    # cedric, owens & perry have e-compasses
     rm /vendor/etc/permissions/android.hardware.sensor.compass.xml
 fi
 
@@ -13,11 +12,8 @@ if [ "$device" == "montana" ]; then
     if [ "$radio" != "APAC" ] && [ "$radio" != "EMEA" ]; then
         # On montana radios other than APAC and EMEA don't have NFC chips
         rm /system/etc/permissions/com.android.nfc_extras.xml
-        rm /vendor/etc/libnfc-nci.conf
-        rm /vendor/etc/libnfc-nxp.conf
         rm /vendor/etc/permissions/android.hardware.nfc.xml
         rm /vendor/etc/permissions/android.hardware.nfc.hce.xml
-        rm /vendor/firmware/libpn548ad_fw.so
         rm -r /system/app/NfcNci
     fi
     if [ "$radio" == "APAC" ] || [ "$radio" == "EMEA" ]; then
